refactor: move the application settings to preference module
Move the ApplicationSettings into the core:preferences module and change its named to
ApplicationPreferences.
